requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
importlib-metadata==6.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
python-engineio==4.10.1
python-socketio==5.11.4
pip==24.0
protobuf==3.20.3
setuptools==40.8.0
SQLAlchemy==2.0.31
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
python-dotenv==1.0.1